requirements.txt
annotated-types==0.7.0
certifi==2025.11.12
charset-normalizer==3.4.4
fpdf==1.7.2
idna==3.11
markdown2==2.5.4
networkx==3.6.1
numpy==2.4.1
pandas==2.3.3
pdfkit==1.0.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
scipy==1.17.0
six==1.17.0
SQLAlchemy==2.0.45
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
wkhtmltopdf==0.2
